[DEV-847] Refactor to hook format in core
